fix: split routes/tenants.rs under 600-line CI cap
Moves role helpers (TenantRole enum, require_session, require_role,
urlencoding_simple, their tests) to routes/tenants/role.rs (153 lines).
Handlers + handler tests stay in routes/tenants/mod.rs (519 lines).
No behaviour change — same public API, all 30 tests pass.
